requirements.txt
flask
crate
python-ldap
